+++ src/components/media/MediaItemBase.svelte
